+++ HISTORY.md
@@ -6,6 +6,7 @@ Release 1.0.6 (WIP)
 - Make output to screen less noisy by default
 - Add output to JSON flat file to support backend
 - Remove draft label from 802.11ax detection
+- Fix listen only/passive mode (--noAP) which listens for any association request on a set channel
 - Add MCS detection for 802.11ac and 802.11ax
 - Add 160 MHz support for 802.11ac and 802.11ax
 - Add feature detection for 802.11ax TWT, Punctured Preamble, HE (ER) Extended Range, UORA (Uplink OFDMA Random Access), and BSR Control (Buffer Status Report)